created a folder for the Newsletter component and moved Newsletter files into it
